perf: impact analysis 添加影响因子

+++ task-impact-analysis/src/main/java/org/njuse17advancedse/taskimpactanalysis/controller/TaskImpactAnalysisController.java
@@ -21,7 +21,7 @@ public class TaskImpactAnalysisController {
   }
 
   @GetMapping(value = "/paper/{paperId}")
-  public int getPaperImpact(
+  public double getPaperImpact(
     @PathVariable Long paperId,
     @RequestBody PaperVo vo
   ) {
